+++ src/pages/beheerder/WerkbonDetail.tsx
@@ -11,6 +11,7 @@ import { Klusacties } from '@/components/werkbon/Klusacties'
 import { Klusinfo } from '@/components/werkbon/Klusinfo'
 import { Werkdocumenten } from '@/components/werkbon/Werkdocumenten'
 import { Klusplanning } from '@/components/werkbon/Klusplanning'
+import { Klusploeg } from '@/components/werkbon/Klusploeg'
 import { Opleverrapport } from '@/components/werkbon/Opleverrapport'
 import { Spinner } from '@/components/ui/Spinner'
 import { Modal } from '@/components/ui/Modal'
@@ -162,6 +163,8 @@ export default function WerkbonDetail() {
         {/* Waar deze klus in de weekplanning staat. Bij een handmatige
             bon is dit de enige plek waar dat te zetten valt. */}
         <Klusplanning werkbon={werkbon} onKlaar={refetch} />
+
+        <Klusploeg werkbon={werkbon} onKlaar={refetch} />
 
         {/* Kantoor kan de werkopdracht en de tekening er alsnog bij
             zetten. Bij een klus uit ClickUp komen ze vanzelf mee, bij